+++ src/Selectbox.privates.ts
@@ -37,7 +37,7 @@ export const getVisibleLabel = <O extends OptionOrValue | string>(
 	labelFormatter?: (selected: O) => NonNullable<ReactNode>
 ): ReactNode => {
 	const selOption =
-		getSelectedOption(options, value) || (placeholder == null && options[0]);
+		getSelectedOption(options, value) ?? (placeholder == null && options[0]);
 	let visibleLabel: ReactNode;
 	if (selOption !== false && selOption != null) {
 		let optionLabel =